Prepare to remove field_s.h
